+++ src/at/int32/sweaty/ui/controls/Toolbar.java
@@ -1,50 +1,19 @@
 package at.int32.sweaty.ui.controls;
 
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-
-import org.eclipse.swt.graphics.Image;
 import org.eclipse.swt.widgets.Composite;
 import org.eclipse.swt.widgets.ToolBar;
 
-import at.int32.sweaty.ui.EventHandler;
-import at.int32.sweaty.ui.Window;
+import at.int32.sweaty.ui.Control;
 
-public class Toolbar extends Widget<org.eclipse.swt.widgets.ToolBar> {
+public class Toolbar extends Widget<ToolBar>{
 
-	protected ToolBar raw;
-	protected java.util.List<ToolbarItem> items = new ArrayList<>();
-
-	public Toolbar(Window window) {
-		super(window);
+	public Toolbar(Control parent) {
+		super(parent);
 	}
 
 	@Override
-	public ToolBar getBaseControl(Composite parent) {
-		raw = parent.getShell().getToolBar();
-
-		return raw;
+	public ToolBar getBaseControl(Composite parent, int style) {
+		return parent.getShell().getToolBar();
 	}
 
-	// public void addWidget(Control control, EventHandler onToolBarItemClicked)
-	// {
-	// ToolbarItem item = new ToolbarItem(this);
-	// items.add(item);
-	//
-	// item.widget(control);
-	// }
-
-	public java.util.List<ToolbarItem> getToolbarItems() {
-		return (List<ToolbarItem>) Collections.unmodifiableCollection(items);
-	}
-
-	public void addButton(String text, Image image, EventHandler clickHandler) {
-		ToolbarItem item = new ToolbarItem(this);
-		items.add(item);
-
-		item.text("Test");
-		item.image(image);
-		// item.
-	}
 }
