--- src/src/at/int32/sweaty/ui/controls/Toolbar.java
@@ -1,19 +1,0 @@
-package at.int32.sweaty.ui.controls;
-
-import org.eclipse.swt.widgets.Composite;
-import org.eclipse.swt.widgets.ToolBar;
-
-import at.int32.sweaty.ui.Control;
-
-public class Toolbar extends Widget<ToolBar>{
-
-	public Toolbar(Control parent) {
-		super(parent);
-	}
-
-	@Override
-	public ToolBar getBaseControl(Composite parent, int style) {
-		return parent.getShell().getToolBar();
-	}
-
-}
